rename module initPlayer to player
