+++ go.mod
@@ -2,6 +2,6 @@ module go-tic-tac-toe
 
 go 1.17
 
-replace go-tic-tac-toe/internal/initPlayer v0.0.0 => ./internal/initPlayer
+replace go-tic-tac-toe/internal/player v0.0.0 => ./internal/player
 
-require go-tic-tac-toe/internal/initPlayer v0.0.0
+require go-tic-tac-toe/internal/player v0.0.0
