--- internal/initPlayer/go.mod
@@ -1,3 +1,0 @@
-module go-tic-tac-toe/internal/initPlayer
-
-go 1.17
