+++ internal/player/go.mod
@@ -1,0 +1,3 @@
+module go-tic-tac-toe/internal/player
+
+go 1.17
